requirements.txt
python-dotenv==0.7.1
graphene-sqlalchemy==2.0
SQLAlchemy==1.1.15
Flask==0.12.2
Flask-GraphQL==1.4.1
psycopg2==2.7.3
